SlashCommands: Remove interaction handoff message
All commands now respond with `msg.respond` instead of
`msg.channel.send` so that we can override the `msg.respond` function
for slash command interactions. Whenever a command sends a response now,
the first call is directed to a new callback that will call
`interaction_respond()`. The `msg.respond` handler is then set back to
the normal one for normal commands so that any further calls to
`msg.respond` will just send the messages in the channel like normal.